Move shared context "with exclude limit tracking"
The shared context "with exclude limit tracking" should be available for
cop plugin authors, who may be calling `exclude_limit` and testing e.g.
`cop.config_to_allow_offenses` in their specs.

After the next release, you will need to change your specs like in 5c4fd2cfb8be53b62e41c86698c0c7c7e434991a:

```diff
 RSpec.describe YourCop, :config do
+  include_context 'with exclude limit tracking'
 ...
-    expect(cop.config_to_allow_offenses).to eq(exclude_limit: { 'Max' => 81 })
+    expect(read_exclude_limit(cop)).to eq('Max' => 81)
```
